[apr-fixes] refactor for non-ube APR
