+++ src/constants/index.ts
@@ -187,6 +187,8 @@ export const DEFAULT_DEADLINE_FROM_NOW = 60 * 20
 // used for rewards deadlines
 export const BIG_INT_SECONDS_IN_WEEK = JSBI.BigInt(60 * 60 * 24 * 7)
 
+export const BIG_INT_SECONDS_IN_YEAR = JSBI.BigInt(60 * 60 * 24 * 365)
+
 export const BIG_INT_ZERO = JSBI.BigInt(0)
 
 // one basis point
