+++ src/pages/Earn/Manage.tsx
@@ -2,6 +2,7 @@ import { cUSD, JSBI } from '@ubeswap/sdk'
 import QuestionHelper from 'components/QuestionHelper'
 import React, { useCallback, useState } from 'react'
 import { Link, RouteComponentProps, useLocation } from 'react-router-dom'
+import { usePairStakingInfo } from 'state/stake/useStakingInfo'
 import styled from 'styled-components'
 import { CountUp } from 'use-count-up'
 
@@ -20,7 +21,7 @@ import { useCurrency } from '../../hooks/Tokens'
 import { useColor } from '../../hooks/useColor'
 import usePrevious from '../../hooks/usePrevious'
 import { useWalletModalToggle } from '../../state/application/hooks'
-import { usePairDualStakingInfo, usePairStakingInfo } from '../../state/stake/hooks'
+import { usePairDualStakingInfo } from '../../state/stake/hooks'
 import { useTokenBalance } from '../../state/wallet/hooks'
 import { ExternalLinkIcon, TYPE } from '../../theme'
 import { currencyId } from '../../utils/currencyId'
