+++ src/state/stake/hooks.ts
@@ -8,7 +8,6 @@ import useCurrentBlockTimestamp from 'hooks/useCurrentBlockTimestamp'
 import { zip } from 'lodash'
 // Hooks
 import { useMemo } from 'react'
-import useCUSDPrice from 'utils/useCUSDPrice'
 
 import ERC_20_INTERFACE from '../../constants/abis/erc20'
 import { STAKING_REWARDS_INTERFACE } from '../../constants/abis/staking-rewards'
@@ -23,6 +22,7 @@ import {
   useSingleContractMultipleData,
 } from '../multicall/hooks'
 import { tryParseAmount } from '../swap/hooks'
+import useStakingInfo from './useStakingInfo'
 import { DualRewardsInfo, useDualStakeRewards } from './useDualStakeRewards'
 
 export const POOF_DUAL_POOL = '0x969D7653ddBAbb42589d73EfBC2051432332A940'
@@ -68,13 +68,8 @@ export interface StakingInfo {
   ) => TokenAmount
   readonly nextPeriodRewards: TokenAmount
   readonly poolInfo: IRawPool
-  readonly dollarRewardPerYear: TokenAmount | undefined
   readonly rewardToken: Token | undefined
   readonly dualRewards: boolean
-}
-
-export const usePairStakingInfo = (pairToFilterBy?: Pair | null, stakingAddress?: string): StakingInfo | undefined => {
-  return useStakingInfo(pairToFilterBy, stakingAddress)[0] ?? undefined
 }
 
 export const usePairDualStakingInfo = (stakingInfo: StakingInfo | undefined): DualRewardsInfo | null => {
@@ -163,169 +158,6 @@ export const useUnclaimedStakingRewards = (): UnclaimedInfo => {
     earned,
     noncirculatingSupply: balanceRemaining && earned ? balanceRemaining.sub(earned) : null,
   }
-}
-
-// gets the staking info from the network for the active chain id
-export function useStakingInfo(pairToFilterBy?: Pair | null, stakingAddress?: string): readonly StakingInfo[] {
-  const { chainId, account } = useActiveWeb3React()
-  const ube = chainId ? UBE[chainId] : undefined
-  const ubePrice = useCUSDPrice(ube)
-
-  // detect if staking is ended
-  const currentBlockTimestamp = useCurrentBlockTimestamp()
-
-  const info = useStakingPools(pairToFilterBy, stakingAddress)
-
-  // These are the staking pools
-  const rewardsAddresses = useMemo(() => info.map(({ stakingRewardAddress }) => stakingRewardAddress), [info])
-
-  const accountArg = useMemo(() => [account ?? undefined], [account])
-
-  // get all the info from the staking rewards contracts
-  const balances = useMultipleContractSingleData(rewardsAddresses, STAKING_REWARDS_INTERFACE, 'balanceOf', accountArg)
-  const earnedAmounts = useMultipleContractSingleData(rewardsAddresses, STAKING_REWARDS_INTERFACE, 'earned', accountArg)
-  const totalSupplies = useMultipleContractSingleData(rewardsAddresses, STAKING_REWARDS_INTERFACE, 'totalSupply')
-
-  // tokens per second, constants
-  const rewardRates = useMultipleContractSingleData(
-    rewardsAddresses,
-    STAKING_REWARDS_INTERFACE,
-    'rewardRate',
-    undefined,
-    NEVER_RELOAD
-  )
-
-  const periodFinishes = useMultipleContractSingleData(
-    rewardsAddresses,
-    STAKING_REWARDS_INTERFACE,
-    'periodFinish',
-    undefined,
-    NEVER_RELOAD
-  )
-
-  return useMemo(() => {
-    if (!chainId || !ube) return []
-
-    return info.reduce(
-      (memo: StakingInfo[], { stakingRewardAddress: rewardsAddress, poolInfo, tokens }, index: number) => {
-        // these two are dependent on account
-        const balanceState = balances[index]
-        const earnedAmountState = earnedAmounts[index]
-
-        // these get fetched regardless of account
-        const totalSupplyState = totalSupplies[index]
-        const rewardRateState = rewardRates[index]
-        const periodFinishState = periodFinishes[index]
-
-        if (
-          // these may be undefined if not logged in
-          !balanceState?.loading &&
-          !earnedAmountState?.loading &&
-          // always need these
-          totalSupplyState &&
-          !totalSupplyState.loading &&
-          rewardRateState &&
-          !rewardRateState.loading &&
-          periodFinishState &&
-          !periodFinishState.loading
-        ) {
-          if (
-            balanceState?.error ||
-            earnedAmountState?.error ||
-            totalSupplyState.error ||
-            rewardRateState.error ||
-            periodFinishState.error
-          ) {
-            console.error('Failed to load staking rewards info')
-            return memo
-          }
-
-          const rewardToken = poolInfo.rewardToken
-            ? new Token(chainId, poolInfo.rewardToken, 18, poolInfo.rewardTokenSymbol)
-            : ube
-
-          // get the LP token
-          const liquidityToken = new Token(chainId, poolInfo.stakingToken, 18, 'ULP', 'Ubeswap LP Token')
-
-          // check for account, if no account set to 0
-          const stakedAmount = new TokenAmount(liquidityToken, JSBI.BigInt(balanceState?.result?.[0] ?? 0))
-          const totalStakedAmount = new TokenAmount(liquidityToken, JSBI.BigInt(totalSupplyState.result?.[0]))
-          const totalRewardRate = new TokenAmount(rewardToken, JSBI.BigInt(rewardRateState.result?.[0]))
-          const nextPeriodRewards = new TokenAmount(ube, poolInfo.nextPeriodRewards?.toString() ?? '0')
-
-          // tokens per month
-          const ubePerYear =
-            rewardToken === ube
-              ? new TokenAmount(ube, JSBI.multiply(totalRewardRate.raw, JSBI.BigInt(365 * 24 * 60 * 60)))
-              : new TokenAmount(ube, '0')
-          const dollarRewardPerYear = ubePrice?.quote(ubePerYear)
-
-          const getHypotheticalRewardRate = (
-            stakedAmount: TokenAmount,
-            totalStakedAmount: TokenAmount,
-            totalRewardRate: TokenAmount
-          ): TokenAmount => {
-            return new TokenAmount(
-              rewardToken,
-              JSBI.greaterThan(totalStakedAmount.raw, JSBI.BigInt(0))
-                ? JSBI.divide(JSBI.multiply(totalRewardRate.raw, stakedAmount.raw), totalStakedAmount.raw)
-                : JSBI.BigInt(0)
-            )
-          }
-
-          const individualRewardRate = getHypotheticalRewardRate(stakedAmount, totalStakedAmount, totalRewardRate)
-
-          const periodFinishSeconds = periodFinishState.result?.[0]?.toNumber()
-          const periodFinishMs = periodFinishSeconds * 1000
-
-          // compare period end timestamp vs current block timestamp (in seconds)
-          const active =
-            periodFinishSeconds && currentBlockTimestamp
-              ? periodFinishSeconds > currentBlockTimestamp.toNumber()
-              : false
-
-          if (!tokens) {
-            return memo
-          }
-
-          memo.push({
-            stakingRewardAddress: rewardsAddress,
-            stakingToken: totalStakedAmount.token,
-            tokens,
-            periodFinish: periodFinishMs > 0 ? new Date(periodFinishMs) : undefined,
-            earnedAmount: new TokenAmount(rewardToken, JSBI.BigInt(earnedAmountState?.result?.[0] ?? 0)),
-            earnedAmountUbe: new TokenAmount(rewardToken, JSBI.BigInt(earnedAmountState?.result?.[0] ?? 0)),
-            rewardRate: individualRewardRate,
-            ubeRewardRate: individualRewardRate,
-            totalRewardRate: totalRewardRate,
-            totalUBERewardRate: totalRewardRate,
-            nextPeriodRewards,
-            stakedAmount: stakedAmount,
-            totalStakedAmount: totalStakedAmount,
-            getHypotheticalRewardRate,
-            active,
-            poolInfo,
-            dollarRewardPerYear,
-            rewardToken,
-            dualRewards: false,
-          })
-        }
-        return memo
-      },
-      []
-    )
-  }, [
-    balances,
-    chainId,
-    currentBlockTimestamp,
-    earnedAmounts,
-    info,
-    periodFinishes,
-    rewardRates,
-    totalSupplies,
-    ube,
-    ubePrice,
-  ])
 }
 
 interface IStakingPool {
